backend/requirements.txt
django==5.2.10
djangorestframework==3.16.1
django-environ==0.12.0
mysqlclient==2.2.7
python-dotenv
gunicorn
drf-yasg==1.21.11
django-cors-headers